+++ src/i8c/constants.py
@@ -36,8 +36,6 @@ I8_TYPE_PTR = "p"
 I8_TYPE_OPAQUE = "o"
 I8_TYPE_FUNC = "F"
 
-I8_EXT_FUNCTION = "f"
-
 DW_OP_addr = 0x03
 DW_OP_deref = 0x06
 DW_OP_const1u = 0x08